requirements.txt
pillow
pytesseract
easyocr
opencv-python
spacy
rapidfuzz